backend/data/project-helios/haas/requirements.txt
# HaaS Platform - Homologação como Serviço
# Requirements básicos para Fase 1
# Data: 14 de outubro de 2025

# ===== CORE API =====
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# ===== DATABASE & STORAGE =====
SQLAlchemy>=2.0.23
alembic>=1.12.0
psycopg2-binary>=2.9.9
GeoAlchemy2>=0.14.0
redis>=5.0.1

# ===== VALIDATION & PROCESSING =====
PyMuPDF>=1.22.5
pypdf>=4.0.0
jsonschema>=4.17.0

# ===== HTTP & INTEGRATIONS =====
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.8.6
tenacity>=8.2.3
beautifulsoup4>=4.12.2
lxml>=4.9.3
psutil>=5.9.5

# ===== DATA PROCESSING =====
pandas>=2.1.1
openpyxl>=3.1.2

# ===== UTILITIES =====
python-dotenv>=1.0.0
pathlib2>=2.3.7

# ===== SECURITY =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
